requirements.txt
torch
torchvision
Pillow
numpy
tqdm
matplotlib
datasets
